fix: category services
