+++ src/components/global/server/command-pallet.tsx
@@ -2,7 +2,7 @@ import "server-only";
 
 import React from "react";
 import { CategoryCommand } from "@/components/global/parts/category-command";
-import { searchCategories } from "@/services/category/get";
+import { searchCategories } from "@/services/category/get/search";
 
 export async function CommandPallet() {
   const categories = await searchCategories(null, 20);
