+++ src/services/category/post/index.ts
@@ -1,9 +1,11 @@
-import { prisma } from "@/lib/client/prisma";
+import { prisma, runPrisma } from "@/lib/client/prisma";
 import { CategoryInput } from "@/lib/schema/server/category";
 
 export const createCategory = async (input: CategoryInput) => {
-  return prisma.category.create({
-    data: input,
-    select: { name: true },
-  });
+  return runPrisma(() =>
+    prisma.category.create({
+      data: input,
+      select: { name: true },
+    })
+  );
 };
